html/browsers/browsing-the-web/navigating-across-documents/008.html

<!doctype html>
<title>Link with onclick form submit to javascript: url is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:parent.postMessage('click', '*')"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>
/*---
specid: navigating-across-documents
steps: [6, 11, 12]
info: |
  Cancel any preexisting but not yet mature attempt to navigate browsingContext, including canceling any instances of the fetch algorithm started by those attempts. If one of those attempts has already created and initialized a new Document object, abort that Document also. (Navigation attempts that have matured already have session history entries, and are therefore handled during the update the session history with the new page algorithm, later.)

  Step 12 contains https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol

  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation

description: |
  1. This test creates a form with `target="test"` and an anchor with both an onclick and href.
  2. The test "clicks" the anchor, which will trigger the onclick operation, which will attempt to submit the form. The form's `action` is a `javascript:` url, which has special handling in step 12
    TODO: Determine if this is supposed to occur in a later task queue.
  3. Navigating to the anchor's `href=""` will abort the `action`.
---*/

var t = async_test();
t.step(function() {
  document.getElementsByTagName("a")[0].click();
});
window.form = document.getElementsByTagName("form")[0];
onmessage = t.step_func(function(e) {
  assert_equals(e.data, "href");
  t.done();
});
</script>
